requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
pypdf==3.9.0
azure-identity==1.13.0b3
azure-search-documents==11.4.0b3
azure-storage-blob==12.14.1
azure-core
pandas
pdfkit
python-dotenv
openpyxl
langchain